Semi-functional double precision

+++ app/src/main/java/pro/shpin/kirill/simplerendering/game/Vector3d.java
@@ -4,10 +4,10 @@ package pro.shpin.kirill.simplerendering.game;
  * Created by wiish on 6/9/2017.
  */
 
-public class Vector3f {
-	public float x, y, z;
+public class Vector3d {
+	public double x, y, z;
 
-	public Vector3f(float x, float y) {
+	public Vector3d(double x, double y) {
 		this.x = x;
 		this.y = y;
 		this.z = 1;
